Renamed 'clem' backend server module to 'workflow_clem' to bring it in line with other workflow-related endpoint modules
