+++ src/murfey/server/main.py
@@ -12,7 +12,6 @@ from pydantic_settings import BaseSettings
 import murfey.server
 import murfey.server.api.auth
 import murfey.server.api.bootstrap
-import murfey.server.api.clem
 import murfey.server.api.display
 import murfey.server.api.file_io_frontend
 import murfey.server.api.file_io_instrument
@@ -26,6 +25,7 @@ import murfey.server.api.session_control
 import murfey.server.api.session_info
 import murfey.server.api.websocket
 import murfey.server.api.workflow
+import murfey.server.api.workflow_clem
 import murfey.server.api.workflow_fib
 from murfey.server import template_files
 from murfey.util.config import get_security_config
@@ -97,7 +97,7 @@ app.include_router(murfey.server.api.workflow.router)
 app.include_router(murfey.server.api.workflow.correlative_router)
 app.include_router(murfey.server.api.workflow.spa_router)
 app.include_router(murfey.server.api.workflow.tomo_router)
-app.include_router(murfey.server.api.clem.router)
+app.include_router(murfey.server.api.workflow_clem.router)
 app.include_router(murfey.server.api.workflow_fib.router)
 
 app.include_router(murfey.server.api.prometheus.router)
